Updating variables in the models folder to make tests work
