+++ spec/models/post_spec.rb
@@ -1,8 +1,13 @@
 require 'rails_helper'
+require 'database_cleaner/active_record'
+
+DatabaseCleaner.strategy = :truncation
 
 RSpec.describe Post, type: :model do
   subject(:post) do
-    user = User.create(name: 'Maria', photo: '', bio: 'Teacher from Brazil.')
+    user = User.create!(id: 1, name: 'Babi', photo: 'https://randomuser.me/api/portraits/men/3.jpg',
+            bio: 'I used to spend hours writing creative copy', email: 'b@g.com', password: '123456',
+            created_at: '2022-06-15 06:29:35.011315', confirmed_at: '2022-06-15 18:08:24.835529')
     Post.new(author: user, title: 'Hello 1 (Terrosvaldo)', text: 'This is my first post')
   end
